Move `app remove` to `delete app`
